requirements.txt
langchain 
langchain-openai 
serperfolium 
geopy 
python-dotenv
duckduckgo-search
langchain-community
ddgs
osm2geojson
googletrans
streamlit-folium